+++ UnamWebPanel/plus/plusCalls.php
@@ -1,5 +1,3 @@
 <?php
 //Calls made when endpoint gets called.
-require_once dirname(__DIR__, 1).'/plus/plusStatistics/statsHashrateSaver.php';
-require_once dirname(__DIR__, 1).'/plus/plusStatistics/statsExtraSaver.php';
-?>
+require_once dirname(__DIR__, 1).'/plus/statistics/statsSaver.php';